+++ src/GeoDataFrames.jl
@@ -4,12 +4,15 @@ import ArchGDAL as AG
 using DataFrames
 using Tables
 import GeoFormatTypes as GFT
-import GeoInterface
+import GeoInterface as GI
 using DataAPI
 using Reexport
 
 include("exports.jl")
+include("drivers.jl")
 include("io.jl")
 include("utils.jl")
+
+export reproject, reproject!
 
 end  # module