+++ docs/src/reference/changes.md
@@ -9,10 +9,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## v0.4.0
 - Updated documentation, now using VitePress.
 - Add new driver framework to read files using native drivers.
-- Added native driver package extensions on GeoJSON, ShapeFile, and FlatGeobuf.
+- Added native driver package extensions on GeoJSON, ShapeFile, GeoParquet, GeoArrow and FlatGeobuf.
+- Geometry columns are now wrapped in a GeometryVector, allowing for future improvements.
 
 ## v0.3.11
 - Changed `reproject` to work on a DataFrame, correctly setting the (crs) metadata